Refactory initializer for IdentityDbContext
